+++ BloodRefill/Plugin.cs
@@ -29,8 +29,10 @@ namespace VMods.BloodRefill
 			}
 			Utils.Initialize(Log, PluginInfo.PLUGIN_NAME);
 
+			CommandSystemConfig.Initialize(Config);
 			BloodRefillConfig.Initialize(Config);
 
+			CommandSystem.Initialize();
 			BloodRefillSystem.Initialize();
 
 			_hooks = Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly());
@@ -44,8 +46,11 @@ namespace VMods.BloodRefill
 			{
 				return true;
 			}
+			VModStorage.SaveAll();
+
 			_hooks?.UnpatchSelf();
 			BloodRefillSystem.Deinitialize();
+			CommandSystem.Deinitialize();
 			Config.Clear();
 			Utils.Deinitialize();
 			return true;